Integrate translator module, render index template, and implement translation routes
Importing the translator module: The line from machinetranslation import translator is added to import the translator module from the machinetranslation package. This assumes that the translator module contains the necessary translation functions (english_to_french and french_to_english).

Rendering the index template: The renderIndexPage() function is modified to return the rendered index.html template using render_template("index.html"). create an HTML template file called index.html and render it when the root route ("/") is accessed.

Translating English to French: In the /englishToFrench route, the textToTranslate parameter is obtained from the query parameters using request.args.get('textToTranslate'). Then, the obtained text is passed to the translator.english_to_french() function to perform the translation. The translated text is returned as the response.

Translating French to English: Similarly, in the /frenchToEnglish route, the textToTranslate parameter is obtained, and the translation is performed using the translator.french_to_english() function. The translated text is returned as the response.